Testcases for TimeDetails
Started it off by writing down all the methods are neccessary for testing. Then, the made the TestCase classes with different inputs for the testcase. One testcase is for normal formating (meaning the user inputs stuff correctly). The other testcase checks to see if the user passed in non-numbers, then can the formatter check?